- Suppression des callbacks dans PrescriptionAI.kt dans analyse() - intégration des listes de traitement dans LoaderFragment.kt - réception des traitements dans ListeTraitementsFragment.kt (mais il n'en fait rien...)
